refactor: move card-data to separate data folder
